+++ src/test/java/org/mmarini/rl/agents/PPOAgentCreateTest.java
@@ -47,7 +47,7 @@ class PPOAgentCreateTest {
     private static final float EPSILON = 1e-6f;
     private static final String AGENT_YAML = """
             ---
-            $schema: https://mmarini.org/wheelly/ppo-agent-spec-schema-0.1
+            $schema: https://mmarini.org/wheelly/ppo-agent-spec-schema-0.2
             class: org.mmarini.rl.agents.PPOAgent
             rewardAlpha: 0.001
             eta: 1e-3
@@ -100,7 +100,7 @@ class PPOAgentCreateTest {
             """;
     private static final String AGENT_NO_ACTION_ALPHAS_YAML = """
             ---
-            $schema: https://mmarini.org/wheelly/ppo-agent-spec-schema-0.1
+            $schema: https://mmarini.org/wheelly/ppo-agent-spec-schema-0.2
             class: org.mmarini.rl.agents.PPOAgent
             rewardAlpha: 0.001
             eta: 1e-3
@@ -143,7 +143,7 @@ class PPOAgentCreateTest {
             """;
     private static final String AGENT_ACTION_CRITIC_YAML = """
             ---
-            $schema: https://mmarini.org/wheelly/ppo-agent-spec-schema-0.1
+            $schema: https://mmarini.org/wheelly/ppo-agent-spec-schema-0.2
             class: org.mmarini.rl.agents.PPOAgent
             rewardAlpha: 0.001
             eta: 1e-3
@@ -198,7 +198,7 @@ class PPOAgentCreateTest {
             """;
     private static final String AGENT_NO_CRITIC_YAML = """
             ---
-            $schema: https://mmarini.org/wheelly/ppo-agent-spec-schema-0.1
+            $schema: https://mmarini.org/wheelly/ppo-agent-spec-schema-0.2
             class: org.mmarini.rl.agents.PPOAgent
             rewardAlpha: 0.001
             eta: 1e-3
@@ -346,14 +346,13 @@ class PPOAgentCreateTest {
         assertEquals(3f, agent.alphas().get("output"));
         assertEquals(0.5f, agent.lambda());
         assertThat(agent.network().state().getBias("layer1"),
-                matrixCloseTo(new float[][]{
-                        {0.5f, 0.5f}
-                }, EPSILON));
+                matrixCloseTo(new long[]{1, 2}, EPSILON,
+                        0.5f, 0.5f
+                ));
 
         assertThat(agent.network().state().getBias("layer1"),
-                matrixCloseTo(new float[][]{
-                        {0.5f, 0.5f}
-                }, EPSILON));
-
+                matrixCloseTo(new long[]{1, 2}, EPSILON,
+                        0.5f, 0.5f
+                ));
     }
 }