+++ src/test/java/org/mmarini/rl/agents/TDAgentSingleNNEnvCreateTest.java
@@ -42,7 +42,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 class TDAgentSingleNNEnvCreateTest {
     private static final String YAML = """
             ---
-            $schema: https://mmarini.org/wheelly/agent-single-nn-schema-0.4
+            $schema: https://mmarini.org/wheelly/agent-single-nn-schema-0.5
             class: org.mmarini.rl.agents.TDAgentSingleNN
             modelPath: models/test
             seed: 1234