+++ CHANGELOG.md
@@ -10,8 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Issue [#442](https://github.com/m-marini/wheellyj/issues/442): Add web cam processing
-- Issue [#448](https://github.com/m-marini/wheellyj/issues/448):  Add labels in polar radar map
-- Issue [#449](https://github.com/m-marini/wheellyj/issues/449):  Add find label state in state machine engine
+- Issue [#448](https://github.com/m-marini/wheellyj/issues/448): Add labels in polar radar map
+- Issue [#449](https://github.com/m-marini/wheellyj/issues/449): Add find label state in state machine engine
+- Issue [#452](https://github.com/m-marini/wheellyj/issues/452): Add label state to polar environment
+
 
 ### Changed